v1.7.3(purchase + sql):
done with purchase area, create stored procedures in mysql for delete data. Update flight area + minor updates

+++ src/main/java/aviatickets/app/flight/entity/Leg.java
@@ -15,7 +15,7 @@ public record Leg(
 		Date arrivalTime,
 		String duration,
 		@Positive
-		Integer distance,
+		Short distance,
 		String status
 
 ) {}
